Sentiment-Analysis-Simple_RNN
"Sentiment-Analysis-Simple_RNN" appears to be a topic related to sentiment analysis using Simple Recurrent Neural Networks (RNNs). Sentiment analysis is a natural language processing (NLP) task that involves determining the sentiment expressed in a piece of text, whether it's positive, negative, or neutral. RNNs are a type of neural network designed to handle sequential data by maintaining a memory of previous inputs.

In the context of sentiment analysis, a Simple RNN can be used to process text data sequentially, taking into account the order of words in a sentence or document. By analyzing the sequential nature of language, Simple RNNs can capture contextual information that may be crucial for determining sentiment.